feat: updater total number of items in HeaderCartButton
total number of items were not updated in the headerCartButton when item was added to the cart. Reduce funciton is used to, as it provides sum of the all the items quantity added to the cart and provides single value. Everytime each item is added to the cart, the total quantity gets updated in the headerCartButton.