feat: remove item from the Cart
logic to remove the item from the cart is implemented in the CartProvider, by updating the state with reducer.
There are two logics implemented:

When item in the cart has only quantity of 1 and it is being removed, then the whole item is removed from the cart.
When item in the cart has quantity more then 1, so only the quantity of the item removed is updated.